RenderAbort, EventMove, -DistroCat, +WorkflowUpdate
Added the ability to abort the rendering process.
Added the ability to move an event from one place in the event queue to another (i.e. to change it's exec time.)
Removed the Distribution Catalog.
Removed Expressions & Evaluator.
Numerous code formatting updates.
Added a sample for workflow update and re-execution.